Corrected bug in sulcal parametrization (wrong type for a variable).
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/cortical_surface/cortical_surface-gpl/branches/bug_fix@79146 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ brainvisa/toolboxes/cortical_surface/processes/anatomy/ParameterizeSulcus2015.py
@@ -541,7 +541,7 @@ def execution( self, context ):
           a[i]=(newCoord[i+1]-newCoord[i])
           b[i]=(newCoord[i]*(i+1)-newCoord[i+1]*i)
      for i in range(N):
-          ind=math.floor(texOut[0][i])
+          ind=int(math.floor(texOut[0][i]))
           isoParam[0][i]=(texOut[0][i]*a[ind]) + b[ind]
           parameter[i]=(texOut[0][i]*a[ind]) + b[ind]
      ws.write( isoParam, self.texture_param1.fullPath() )
